Wed, Feb 8, 2023, 10:48 PM +00:00

+++ cs50/readability/readability.c
@@ -17,7 +17,7 @@ int main(void)
     int sent = count_sentences(all);
 
     float L = ((letters/(float)words) * 100);
-    float S = ((sent/(float)words) * (int) 100);
+    float S = ((sent/(float)words) * 100);
